Update and rename phenotype.rst to biocharacteristic.rst
The use of phenotype is extremely ambiguous, as demonstrated in the description itself which extends and clarifies the scope of the term's use, in this context.

Additionally, the object is used in connotations such as `biosample`, where the use of "phenotype" seems stretched.

Here, I propose to change the object's and description to a more neutral "biocharacteristic" - as, in fact, it had been discussed and developed in the GA4GH metadata schema, with full awareness of the "phenotype", "disease" etc. use cases.

Therefore, a particular **phenotype** could be described through one or more "biocharacteristics"; but the same object could be used beyond what we would consider a "phenotype".

Also - any other name is fine, as long as it does not refer to a "phenotype" / "phenotypic feature" per se - the same structures etc. can easily be used for histologies etc.